Fix build warnings and update CI configuration
- Android: Guard set_prio function with !PJ_ANDROID to fix unused function warning
- config_site_test.h: Make DTLS conditional on OpenSSL (PJ_SSL_SOCK_IMP_OPENSSL)
- ci-mac.yml: Remove duplicate /lib from pkg-config lib paths
- conf_thread.c: Add parentheses around && within || to fix precedence warning
- ffmpeg_vid_codecs.c: Rename shadowing variable fmt_id to pix_fmt_id, add const to AVCodec pointers, initialize got_picture to 0
- ffmpeg_dev.c: Add const to AVInputFormat pointer, mark unused buf variable
- ssl_sock_gtls.c: Add explicit cast to long for time_t conversions

+++ pjlib/src/pj/ssl_sock_gtls.c
@@ -974,8 +974,8 @@ static void tls_cert_get_info(pj_pool_t *pool, pj_ssl_cert_info *ci,
     tls_cert_get_cn(&ci->subject.info, &ci->subject.cn);
 
     /* Validity */
-    ci->validity.end.sec = gnutls_x509_crt_get_expiration_time(cert);
-    ci->validity.start.sec = gnutls_x509_crt_get_activation_time(cert);
+    ci->validity.end.sec = (long)gnutls_x509_crt_get_expiration_time(cert);
+    ci->validity.start.sec = (long)gnutls_x509_crt_get_activation_time(cert);
     ci->validity.gmt = 0;
 
     /* Subject Alternative Name extension */
